+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/presentation/di/PresentationModule.kt
@@ -2,10 +2,12 @@ package dev.rikoapp.cleanphonelauncher.presentation.di
 
 import dev.rikoapp.cleanphonelauncher.presentation.applist.AppListViewModel
 import dev.rikoapp.cleanphonelauncher.presentation.home.HomeViewModel
+import dev.rikoapp.cleanphonelauncher.presentation.settings.SettingsViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val presentationModule = module {
     viewModelOf(::HomeViewModel)
     viewModelOf(::AppListViewModel)
+    viewModelOf(::SettingsViewModel)
 }